+++ examples/src/tests/bessel.rs
@@ -3,7 +3,7 @@ use autd3::{driver::link::Link, prelude::*};
 pub async fn bessel(autd: &mut Controller<impl Link>) -> anyhow::Result<bool> {
     autd.send(Silencer::default()).await?;
 
-    let center = autd.geometry().center();
+    let center = autd.center();
     let dir = Vector3::z_axis();
 
     let g = Bessel::new(center, dir, 18. / 180. * PI * rad);